+++ lib/src/item_scroll_physics.dart
@@ -1,0 +1,68 @@
+import 'dart:math';
+
+import 'package:flutter/material.dart';
+
+class ItemScrollPhysics extends ScrollPhysics {
+  /// Creates physics for snapping to item.
+  /// Based on PageScrollPhysics
+  final double? itemHeight;
+  final double targetPixelsLimit;
+
+  const ItemScrollPhysics({
+    ScrollPhysics? parent,
+    this.itemHeight,
+    this.targetPixelsLimit = 3.0,
+  })  : assert(itemHeight != null && itemHeight > 0),
+        super(parent: parent);
+
+  @override
+  ItemScrollPhysics applyTo(ScrollPhysics? ancestor) {
+    return ItemScrollPhysics(
+        parent: buildParent(ancestor), itemHeight: itemHeight);
+  }
+
+  double _getItem(ScrollPosition position) {
+    double maxScrollItem =
+        (position.maxScrollExtent / itemHeight!).floorToDouble();
+    return min(max(0, position.pixels / itemHeight!), maxScrollItem);
+  }
+
+  double _getPixels(ScrollPosition position, double item) {
+    return item * itemHeight!;
+  }
+
+  double _getTargetPixels(
+    ScrollPosition position,
+    Tolerance tolerance,
+    double velocity,
+  ) {
+    double item = _getItem(position);
+    if (velocity < -tolerance.velocity) {
+      item -= targetPixelsLimit;
+    } else if (velocity > tolerance.velocity) {
+      item += targetPixelsLimit;
+    }
+    return _getPixels(position, item.roundToDouble());
+  }
+
+  @override
+  Simulation? createBallisticSimulation(
+      ScrollMetrics position, double velocity) {
+    // If we're out of range and not headed back in range, defer to the parent
+    // ballistics, which should put us back in range at a item boundary.
+//    if ((velocity <= 0.0 && position.pixels <= position.minScrollExtent) ||
+//        (velocity >= 0.0 && position.pixels >= position.maxScrollExtent))
+//      return super.createBallisticSimulation(position, velocity);
+    Tolerance tolerance = this.tolerance;
+    final double target =
+        _getTargetPixels(position as ScrollPosition, tolerance, velocity);
+    if (target != position.pixels) {
+      return ScrollSpringSimulation(spring, position.pixels, target, velocity,
+          tolerance: tolerance);
+    }
+    return null;
+  }
+
+  @override
+  bool get allowImplicitScrolling => false;
+}